prepare Layout component data in a common/ directory so it can be reused server-side
